requirements.txt
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.3
Django==3.1.2
django-bootstrap-modal-forms==2.0.0
django-phonenumber-field==5.0.0
django-richtextfield==1.6
django-sass==1.0.0
docopt==0.6.2
docutils==0.16
git-lint==0.1.2
html-linter==0.4.0
isort==5.5.4
lazy-object-proxy==1.4.3
libsass==0.20.1
logilab-common==1.7.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathlib2==2.3.5
pep8==1.7.1
phonenumberslite==8.12.10
Pillow==7.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.3.1
template-remover==0.1.9
termcolor==1.1.0
toml==0.10.1
typing-extensions==3.7.4.3
wrapt==1.12.1
